Extract template list to constants
