+++ tests/Twig/Extension/RenderElementExtensionTest.php
@@ -19,6 +19,7 @@ use Sonata\AdminBundle\Admin\AdminInterface;
 use Sonata\AdminBundle\FieldDescription\FieldDescriptionInterface;
 use Sonata\AdminBundle\SonataConfiguration;
 use Sonata\AdminBundle\Templating\MutableTemplateRegistryInterface;
+use Sonata\AdminBundle\Templating\TemplateRegistryInterface;
 use Sonata\AdminBundle\Tests\Fixtures\Entity\FooToString;
 use Sonata\AdminBundle\Tests\Fixtures\StubFilesystemLoader;
 use Sonata\AdminBundle\Twig\Extension\RenderElementExtension;
@@ -234,39 +235,7 @@ final class RenderElementExtensionTest extends TestCase
         $this->fieldDescription
             ->method('getTemplate')
             ->willReturnCallback(static function () use ($type): ?string {
-                switch ($type) {
-                    case FieldDescriptionInterface::TYPE_STRING:
-                        return '@SonataAdmin/CRUD/list_string.html.twig';
-                    case FieldDescriptionInterface::TYPE_BOOLEAN:
-                        return '@SonataAdmin/CRUD/list_boolean.html.twig';
-                    case FieldDescriptionInterface::TYPE_DATETIME:
-                        return '@SonataAdmin/CRUD/list_datetime.html.twig';
-                    case FieldDescriptionInterface::TYPE_DATE:
-                        return '@SonataAdmin/CRUD/list_date.html.twig';
-                    case FieldDescriptionInterface::TYPE_TIME:
-                        return '@SonataAdmin/CRUD/list_time.html.twig';
-                    case FieldDescriptionInterface::TYPE_CURRENCY:
-                        return '@SonataAdmin/CRUD/list_currency.html.twig';
-                    case FieldDescriptionInterface::TYPE_PERCENT:
-                        return '@SonataAdmin/CRUD/list_percent.html.twig';
-                    case FieldDescriptionInterface::TYPE_EMAIL:
-                        return '@SonataAdmin/CRUD/list_email.html.twig';
-                    case FieldDescriptionInterface::TYPE_CHOICE:
-                        return '@SonataAdmin/CRUD/list_choice.html.twig';
-                    case FieldDescriptionInterface::TYPE_ARRAY:
-                        return '@SonataAdmin/CRUD/list_array.html.twig';
-                    case FieldDescriptionInterface::TYPE_TRANS:
-                        return '@SonataAdmin/CRUD/list_trans.html.twig';
-                    case FieldDescriptionInterface::TYPE_URL:
-                        return '@SonataAdmin/CRUD/list_url.html.twig';
-                    case FieldDescriptionInterface::TYPE_HTML:
-                        return '@SonataAdmin/CRUD/list_html.html.twig';
-                    case 'nonexistent':
-                        // template doesn`t exist
-                        return '@SonataAdmin/CRUD/list_nonexistent_template.html.twig';
-                    default:
-                        return null;
-                }
+                return TemplateRegistryInterface::LIST_TEMPLATES[$type] ?? null;
             });
 
         static::assertSame(
@@ -368,34 +337,7 @@ EOT
         $this->fieldDescription
             ->method('getTemplate')
             ->willReturnCallback(static function () use ($type): ?string {
-                switch ($type) {
-                    case FieldDescriptionInterface::TYPE_BOOLEAN:
-                        return '@SonataAdmin/CRUD/show_boolean.html.twig';
-                    case FieldDescriptionInterface::TYPE_DATETIME:
-                        return '@SonataAdmin/CRUD/show_datetime.html.twig';
-                    case FieldDescriptionInterface::TYPE_DATE:
-                        return '@SonataAdmin/CRUD/show_date.html.twig';
-                    case FieldDescriptionInterface::TYPE_TIME:
-                        return '@SonataAdmin/CRUD/show_time.html.twig';
-                    case FieldDescriptionInterface::TYPE_CURRENCY:
-                        return '@SonataAdmin/CRUD/show_currency.html.twig';
-                    case FieldDescriptionInterface::TYPE_PERCENT:
-                        return '@SonataAdmin/CRUD/show_percent.html.twig';
-                    case FieldDescriptionInterface::TYPE_EMAIL:
-                        return '@SonataAdmin/CRUD/show_email.html.twig';
-                    case FieldDescriptionInterface::TYPE_CHOICE:
-                        return '@SonataAdmin/CRUD/show_choice.html.twig';
-                    case FieldDescriptionInterface::TYPE_ARRAY:
-                        return '@SonataAdmin/CRUD/show_array.html.twig';
-                    case FieldDescriptionInterface::TYPE_TRANS:
-                        return '@SonataAdmin/CRUD/show_trans.html.twig';
-                    case FieldDescriptionInterface::TYPE_URL:
-                        return '@SonataAdmin/CRUD/show_url.html.twig';
-                    case FieldDescriptionInterface::TYPE_HTML:
-                        return '@SonataAdmin/CRUD/show_html.html.twig';
-                    default:
-                        return null;
-                }
+                return TemplateRegistryInterface::SHOW_TEMPLATES[$type] ?? null;
             });
 
         static::assertSame(
@@ -443,34 +385,7 @@ EOT
                     return $options['template'];
                 }
 
-                switch ($type) {
-                    case FieldDescriptionInterface::TYPE_BOOLEAN:
-                        return '@SonataAdmin/CRUD/show_boolean.html.twig';
-                    case FieldDescriptionInterface::TYPE_DATETIME:
-                        return '@SonataAdmin/CRUD/show_datetime.html.twig';
-                    case FieldDescriptionInterface::TYPE_DATE:
-                        return '@SonataAdmin/CRUD/show_date.html.twig';
-                    case FieldDescriptionInterface::TYPE_TIME:
-                        return '@SonataAdmin/CRUD/show_time.html.twig';
-                    case FieldDescriptionInterface::TYPE_CURRENCY:
-                        return '@SonataAdmin/CRUD/show_currency.html.twig';
-                    case FieldDescriptionInterface::TYPE_PERCENT:
-                        return '@SonataAdmin/CRUD/show_percent.html.twig';
-                    case FieldDescriptionInterface::TYPE_EMAIL:
-                        return '@SonataAdmin/CRUD/show_email.html.twig';
-                    case FieldDescriptionInterface::TYPE_CHOICE:
-                        return '@SonataAdmin/CRUD/show_choice.html.twig';
-                    case FieldDescriptionInterface::TYPE_ARRAY:
-                        return '@SonataAdmin/CRUD/show_array.html.twig';
-                    case FieldDescriptionInterface::TYPE_TRANS:
-                        return '@SonataAdmin/CRUD/show_trans.html.twig';
-                    case FieldDescriptionInterface::TYPE_URL:
-                        return '@SonataAdmin/CRUD/show_url.html.twig';
-                    case FieldDescriptionInterface::TYPE_HTML:
-                        return '@SonataAdmin/CRUD/show_html.html.twig';
-                    default:
-                        return null;
-                }
+                return TemplateRegistryInterface::SHOW_TEMPLATES[$type] ?? null;
             });
 
         $this->object->name = 'SonataAdmin';
